[Feature]  當滑動 scrollView 時，會同步改變 KeyTimeView 的 SelectedRangeView 位置 1. 在 viewModel 新增 updateTimeFromScrollOffset 處理滑動時的計算
[Fix] 修正在滑動 scrollView 時，KeyTimeView 的 SelectedRangeView 同步移動時會有卡卡的現象
原因：因為時間計算都是以 Int，導致 UI 變化沒有連續性
修正：時間計算與 Model 裡跟時間有關的型別都改成 CGFloat
備註：但是點選 keyTime 按鈕時，時間還是以 Int 為主，因為時間跳耀點不會有浮點數

+++ PicCollage_HW/MusicStateModel.swift
@@ -8,16 +8,16 @@
 import Foundation
 
 struct MusicStateModel {
-    var totalDuration: Int
-    var currentTime: Int
-    var keyTimes: [Int]
+    var totalDuration: CGFloat
+    var currentTime: CGFloat
+    var keyTimes: [CGFloat]
     var selectedRange: TrimmerRangeModel
-    
-    mutating func updateTotalDurarion(value: Int) {
+
+    mutating func updateTotalDurarion(value: CGFloat) {
         totalDuration = value
     }
-    
-    mutating func updateTrimmerRange(by: Int) {
+
+    mutating func updateTrimmerRange(by: CGFloat) {
         let oldStart = selectedRange.start
         let newRange = selectedRange.shift(by: by, totalDuration: totalDuration)
         selectedRange = newRange
